Add a notion of subdir to the filesystem path.
The public functions, create, fetch, and delete take a root directory as an argument,
e.g. "/var/blobs".

The internal functions add a subdir, either "key" or "digest", which is appended to root.
So the top-level directory for the public keys returned to the client are,
for example, "/var/blobs/key/" and the relative directory for a key "abc" is "a/b/c".

The full Unix pathname for the above example would be "/var/blobs/key/a/b/c/blob".

The "digest" subdir is a parallel directory structure, but instead of GUUIDs (as returned
for the public keys), the files under root/digest are organized by a SHA-256 hash of
the file itself.